requirements.txt
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-Seeder==1.2.0